pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emotion-detection-action"
version = "0.1.0"
description = "Human emotion detection SDK for robotics using Vision-Language-Action models"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Emotion Detection Action Contributors"}
]
keywords = ["emotion", "detection", "robotics", "vla", "vision", "language", "action"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "transformers>=4.35.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "sounddevice>=0.4.6",
    "scipy>=1.11.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "mediapipe>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
vla = [
    "accelerate>=0.24.0",
    "bitsandbytes>=0.41.0",
]
serial = [
    "pyserial>=3.5",
]
websocket = [
    "websocket-client>=1.6.0",
]
robot = [
    "pyserial>=3.5",
    "websocket-client>=1.6.0",
]

[project.urls]
Homepage = "https://github.com/saurabh947/emotion-detector"
Documentation = "https://github.com/saurabh947/emotion-detector#readme"
Repository = "https://github.com/saurabh947/emotion-detector"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
emotion_detection_action = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    # Suppress the expected GRU warning during batch-size > 1 unit tests.
    "ignore:.*use_temporal=True is ignored for batch_size.*:UserWarning",
    # Suppress the multi-worker InferenceWorker advisory during tests.
    "ignore:.*InferenceWorker.*num_workers.*:UserWarning",
]

